[REFACTOR] Move origin encryption keys to their own module
