+++ components/core/src/crypto/keys/box_key_pair.rs
@@ -12,8 +12,7 @@ use super::{super::{ANONYMOUS_BOX_FORMAT_VERSION,
             KeyPair,
             KeyRevision,
             KeyType};
-use crate::{crypto::keys::{Key,
-                           NamedRevision},
+use crate::{crypto::keys::NamedRevision,
             error::{Error,
                     Result}};
 use serde_derive::{Deserialize,
@@ -27,8 +26,7 @@ use sodiumoxide::crypto::{box_::{self,
 use std::{borrow::Cow,
           convert::TryFrom,
           fmt,
-          path::{Path,
-                 PathBuf},
+          path::Path,
           str,
           str::FromStr};
 
@@ -51,6 +49,14 @@ pub struct AnonymousBox {
     sender:     NamedRevision,
     /// The encoded ciphertext of the message.
     ciphertext: Vec<u8>,
+}
+
+impl AnonymousBox {
+    pub fn new(sender: NamedRevision, ciphertext: Vec<u8>) -> Self { Self { sender, ciphertext } }
+
+    pub fn sender(&self) -> &NamedRevision { &self.sender }
+
+    pub fn ciphertext(&self) -> &[u8] { &self.ciphertext }
 }
 
 impl fmt::Display for AnonymousBox {
@@ -345,112 +351,12 @@ impl<'a, 'b: 'a> From<&'b str> for WrappedSealedBox<'a> {
 }
 
 ////////////////////////////////////////////////////////////////////////
-// Origin Encryption Key Pair
-////////////////////////////////////////////////////////////////////////
-
-/// Given the name of an origin, generate a new encryption key pair.
-///
-/// The resulting keys will need to be saved to a cache in order to
-/// persist.
-pub fn generate_origin_encryption_key_pair(
-    origin_name: &str)
-    -> (OriginPublicEncryptionKey, OriginSecretEncryptionKey) {
-    let revision = KeyRevision::new();
-    let (pk, sk) = box_::gen_keypair();
-
-    let public = OriginPublicEncryptionKey::from_raw(origin_name.to_string(), revision.clone(), pk);
-    let secret = OriginSecretEncryptionKey::from_raw(origin_name.to_string(), revision.clone(), sk);
-    (public, secret)
-}
-
-////////////////////////////////////////////////////////////////////////
-
-pub struct OriginPublicEncryptionKey {
-    named_revision: NamedRevision,
-    key:            BoxPublicKey,
-    path:           PathBuf,
-}
-
-impl Key for OriginPublicEncryptionKey {
-    type Crypto = BoxPublicKey;
-
-    const EXTENSION: &'static str = "pub";
-    const PERMISSIONS: crate::fs::Permissions = crate::fs::DEFAULT_PUBLIC_KEY_PERMISSIONS;
-    const VERSION_STRING: &'static str = PUBLIC_BOX_KEY_VERSION;
-
-    fn key(&self) -> &BoxPublicKey { &self.key }
-
-    fn named_revision(&self) -> &NamedRevision { &self.named_revision }
-}
-
-from_str_impl_for_key!(OriginPublicEncryptionKey);
-
-try_from_path_buf_impl_for_key!(OriginPublicEncryptionKey);
-
-as_ref_path_impl_for_key!(OriginPublicEncryptionKey);
-
-impl OriginPublicEncryptionKey {
-    pub(crate) fn from_raw(name: String, revision: KeyRevision, key: BoxPublicKey) -> Self {
-        let named_revision = NamedRevision::new(name, revision);
-        let path = named_revision.filename::<Self>();
-        Self { named_revision,
-               key,
-               path }
-    }
-
-    // corresponds to old encrypt_anonymous_box
-    /// Encrypt a secret
-    pub fn encrypt(&self, data: &[u8]) -> WrappedSealedBox {
-        let ciphertext = sealedbox::seal(data, self.key());
-        let anon = AnonymousBox { sender: self.named_revision.clone(),
-                                  ciphertext };
-        // TODO (CM): Eventually do away with WrappedSealedBox; this
-        // is just for compatibility now
-        WrappedSealedBox::from(EncryptedSecret::Anonymous(anon))
-    }
-}
-
-////////////////////////////////////////////////////////////////////////
-
-pub struct OriginSecretEncryptionKey {
-    named_revision: NamedRevision,
-    key:            BoxSecretKey,
-    path:           PathBuf,
-}
-
-impl Key for OriginSecretEncryptionKey {
-    type Crypto = BoxSecretKey;
-
-    const EXTENSION: &'static str = "box.key";
-    const PERMISSIONS: crate::fs::Permissions = crate::fs::DEFAULT_SECRET_KEY_PERMISSIONS;
-    const VERSION_STRING: &'static str = SECRET_BOX_KEY_VERSION;
-
-    fn key(&self) -> &BoxSecretKey { &self.key }
-
-    fn named_revision(&self) -> &NamedRevision { &self.named_revision }
-}
-
-from_str_impl_for_key!(OriginSecretEncryptionKey);
-
-try_from_path_buf_impl_for_key!(OriginSecretEncryptionKey);
-
-as_ref_path_impl_for_key!(OriginSecretEncryptionKey);
-
-impl OriginSecretEncryptionKey {
-    pub(crate) fn from_raw(name: String, revision: KeyRevision, key: BoxSecretKey) -> Self {
-        let named_revision = NamedRevision::new(name, revision);
-        let path = named_revision.filename::<Self>();
-        Self { named_revision,
-               key,
-               path }
-    }
-}
-
-////////////////////////////////////////////////////////////////////////
 // OLD STUFF BELOW
 ////////////////////////////////////////////////////////////////////////
 
 impl BoxKeyPair {
+    #[deprecated(note = "Please use \
+                         habitat_core::crypto::keys::generate_service_encryption_key_pair instead")]
     pub fn generate_pair_for_service<S1, S2>(org: S1, service_group: S2) -> Result<Self>
         where S1: AsRef<str>,
               S2: AsRef<str>
@@ -751,8 +657,12 @@ mod test {
     use super::{super::super::test_support::*,
                 BoxKeyPair,
                 *};
-    use crate::crypto::keys::{generate_service_encryption_key_pair,
+    use crate::crypto::keys::{generate_origin_encryption_key_pair,
+                              generate_service_encryption_key_pair,
                               generate_user_encryption_key_pair,
+                              Key,
+                              OriginPublicEncryptionKey,
+                              OriginSecretEncryptionKey,
                               ServicePublicEncryptionKey,
                               ServiceSecretEncryptionKey,
                               UserPublicEncryptionKey,
